requirements.txt
Django==5.2.1
Pillow==10.2.0
django-tinymce==3.7.1
python-dotenv==1.0.1 